WIP: Modify the kubernetes.yml playbook to allow preparing for offline
This playbook is pseudo-code! just pushing for now

The playbook will make sure that all the required components documented
in https://kubespray.io/#/docs/offline-environment are up before
the cluster bootstraps.

Namely; it will take care of setting up:

Internal repo for deb packages,  internal web server for all the static
binaries,  and finally it will pre-seed all the nodes with all the
container images for the base system.

This gives us a fully supported and documented way to do offline
kubespray.

See ./inventory/offline/group-vars/k8s-cluster.yml for an example for
configuring the nodes to point to our offline package. This is
literally a copy of  https://kubespray.io/#/docs/offline-environment
which means we can defer much of docs.wire.com to official
documentation, reducing burden of maintenance on us