ci.sh: don't use databases-ephemeral
We only install databases-ephemeral in prod to install redis-ephemeral;
and disable cassandra and elasticsearch in the chart-values. This is a
very convoluted way to install redis. Just install the chart directly.

Also; this would cause all the service names to be prefixed with
"databases-ephemeral-" which requires
`charts/redis-ephemeral/templates/helpers.tpl` to strip it again.