[minio] Integrate recent changes to the role
The initial version of the role that was used for the offline package had certain
shortcomings, thus a hot fix was introduced:

  https://github.com/wireapp/ansible-minio/commit/0c239cbce486fbcb3767f833c374549984d3e751

But since then, upstream was integrated that already introduced the ability of
setting download URLs outside of the role. Upstream as well as and additional
refactoring to our two layouts (2 mini-server on one node) approach has landed in
the role which resulted in a new major version

  https://github.com/wireapp/ansible-minio/releases/tag/v2.0.0

This change set updates the pointer to the submodules and integrates recent API
changes of the role into the offline package.

Please note that certain values have been removed or added due to role-internal
defaults that were changed in the work mentioned above.